+++ src/domain/builder/blog/recent-posts.builder.ts
@@ -6,16 +6,18 @@ import { fillPostMetadata } from "./post.filler";
 export const recentPostsBuilder = (cache: ICache) => {
   const config = cache.get("config");
   const posts = cache.get("posts");
-  const { recentPost: recentPostTemplate } = cache.get("templates");
+  const { recentPost } = cache.get("templates");
 
-  if (!recentPostTemplate) {
+  if (!recentPost) {
     throw new Error("There is no template for recent posts.");
   }
+
+  const rencentPostTemplate = recentPost.raw.toString();
 
   return posts
     .map((post: IFile) => getPostMetadata(config, post))
     .sort(postSorter)
     .slice(0, config.recentPosts)
-    .map((pm: IPostMetadata) => fillPostMetadata(recentPostTemplate, pm))
+    .map((pm: IPostMetadata) => fillPostMetadata(rencentPostTemplate, pm))
     .join("");
 };