+++ src/domain/loader/loader.ts
@@ -1,6 +1,6 @@
 import { join } from "path";
 
-import { ICache } from "../../interfaces";
+import { ICache, IFile } from "../../interfaces";
 import { DirScanner } from "../../lib/dir-scanner";
 import { readFile } from "../../lib/io";
 
@@ -30,14 +30,13 @@ export class Loader {
 
   private async loadTemplates() {
     const config = this.cache.get("config");
-    const templates: { [key: string]: string } = {};
+    const templates: { [key: string]: IFile } = {};
     const templateFolder = join(config.projectFolder, config.templateFolder);
     try {
       const files = await DirScanner.scanAndGetFiles(templateFolder);
 
       for (const file of files) {
-        const name = file.name.split(".")[0];
-        templates[name] = file.raw.toString();
+        templates[file.name] = file;
       }
 
       return templates;