+++ test/unit/builder.html.test.ts
@@ -1,11 +1,20 @@
 import { MemoryCache } from "../../src/lib/cache";
-import { IFile } from "../../src/interfaces";
+import { IFile, ICache } from "../../src/interfaces";
 import { buildHtml } from "../../src/domain/builder";
 import { DirScanner } from "../../src/lib/dir-scanner";
 
 describe("Builder - HTML", () => {
+  const defaultConfig = {
+    postsFolder: "posts"
+  };
+  let cache: ICache;
+
+  beforeEach(() => {
+    cache = new MemoryCache();
+  });
+
   test("Builds file without template", async () => {
-    const cache = new MemoryCache();
+    cache.set("config", defaultConfig);
     const fakeFile: IFile = {
       name: "article",
       extension: "html",
@@ -24,7 +33,7 @@ describe("Builder - HTML", () => {
   });
 
   test("Template not found", async () => {
-    const cache = new MemoryCache();
+    cache.set("config", defaultConfig);
     cache.set("templates", {});
     const fakeFile: IFile = {
       name: "article",
@@ -40,9 +49,15 @@ describe("Builder - HTML", () => {
   });
 
   test("Replace content correctly", async () => {
-    const cache = new MemoryCache();
+    cache.set("config", defaultConfig);
     cache.set("templates", {
-      main: "<div>{{content}}</div>"
+      main: {
+        name: "main",
+        extension: "html",
+        path: "",
+        modifiedAt: new Date(),
+        raw: "<div>{{content}}</div>"
+      }
     });
 
     const fakeFile: IFile = {
@@ -62,40 +77,18 @@ describe("Builder - HTML", () => {
     expect(output).toMatchObject(result);
   });
 
-  test("Replace snippets inside template correctly", async () => {
-    const cache = new MemoryCache();
-    cache.set("config", { mode: "prod" });
-    cache.set("templates", {
-      main: "<div>{{snippet:myvar1}}{{content}}</div>"
-    });
-    cache.set("snippets", {
-      myvar1: { configMode: "all", value: "ABCD" },
-      myvar2: { configMode: "all", value: "XYZA" }
-    });
-
-    const fakeFile: IFile = {
-      name: "article",
-      extension: "html",
-      modifiedAt: new Date(),
-      path: "",
-      raw: "<!--\n @template main \n--> <p>{{snippet:myvar2}}</p>"
-    };
-
-    const result: IFile = {
-      ...fakeFile,
-      raw: "<div>ABCD <p>XYZA</p></div>"
-    };
-
-    const output = await buildHtml(cache, fakeFile);
-    expect(output).toMatchObject(result);
-  });
-
   test("Replace snippets configMode", async () => {
-    const cache = new MemoryCache();
-    cache.set("config", { mode: "prod" });
+    cache.set("config", { ...defaultConfig, mode: "prod" });
     cache.set("templates", {
-      main: "<div>{{snippet:myvar1}}{{content}}{{snippet:myvar3}}</div>"
+      main: {
+        name: "main",
+        extension: "html",
+        path: "",
+        modifiedAt: new Date(),
+        raw: "<div>{{snippet:myvar1}}{{content}}{{snippet:myvar3}}</div>"
+      }
     });
+
     cache.set("snippets", {
       myvar1: { configMode: "prod", value: "ABCD" },
       myvar2: { configMode: "dev", value: "XYZA" },
@@ -119,11 +112,24 @@ describe("Builder - HTML", () => {
     expect(output).toMatchObject(result);
   });
 
+  test("Post has no date", () => {
+    // TOODO
+  });
+
+  test("Inject post metadata", () => {
+    // TOODO
+  });
+
   test("No recent post template file", async () => {
-    const cache = new MemoryCache();
-    cache.set("config", { recentPosts: 2 });
+    cache.set("config", { ...defaultConfig, recentPosts: 2 });
     cache.set("templates", {
-      main: "<div>{{content}}</div>"
+      main: {
+        name: "main",
+        extension: "html",
+        path: "",
+        modifiedAt: new Date(),
+        raw: "<div>{{content}}</div>"
+      }
     });
     cache.set("posts", DirScanner.scanAndGetFiles("./test/data/project/posts"));
 
@@ -141,18 +147,27 @@ describe("Builder - HTML", () => {
   });
 
   test("Recent posts", async () => {
-    const cache = new MemoryCache();
-    cache.set("config", { recentPosts: 2, postsFolder: "posts" });
+    cache.set("config", { ...defaultConfig, recentPosts: 2 });
     cache.set("templates", {
-      main: "<div>{{content}}</div>",
-      recentPost:
-        '<div href="{{href}}"><p>{{title}}</p><p>{{author}}</p><p>{{description}}</p><p>{{date}}</p></div>'
+      main: {
+        name: "main",
+        extension: "html",
+        path: "",
+        modifiedAt: new Date(),
+        raw: "<div>{{content}}</div>"
+      },
+      recentPost: {
+        name: "recentPost",
+        extension: "html",
+        path: "",
+        modifiedAt: new Date(),
+        raw:
+          '<div href="{{href}}"><p>{{title}}</p><p>{{author}}</p><p>{{description}}</p><p>{{date}}</p></div>'
+      }
     });
     cache.set(
       "posts",
-      (await DirScanner.scanAndGetFiles("./test/data/project/posts")).map(
-        f => ({ ...f, path: "/some" })
-      )
+      await DirScanner.scanAndGetFiles("./test/data/project/posts")
     );
 
     const fakeFile: IFile = {
@@ -164,9 +179,10 @@ describe("Builder - HTML", () => {
         '<!--\n @template main \n--> <div class="recent-posts">{{blog:recent-posts}}</div>'
     };
 
+    // TODO Check post in subfolder href
     const posts = [
-      '<div href="posts/some/fourth.html"><p>Fourth post</p><p>Test</p><p>Fourth blog test</p><p>04-01-2019</p></div>',
-      '<div href="posts/some/third.html"><p>Third post</p><p>Test</p><p>Third blog test</p><p>03-01-2019</p></div>'
+      '<div href="posts/fourth.html"><p>Fourth post</p><p>Test</p><p>Fourth blog test</p><p>04-01-2019</p></div>',
+      '<div href="posts/third.html"><p>Third post</p><p>Test</p><p>Third blog test</p><p>03-01-2019</p></div>'
     ];
 
     const result: IFile = {
@@ -179,11 +195,22 @@ describe("Builder - HTML", () => {
   });
 
   test("Recent posts - use partial metadata", async () => {
-    const cache = new MemoryCache();
-    cache.set("config", { recentPosts: 2, postsFolder: "posts" });
+    cache.set("config", { ...defaultConfig, recentPosts: 2 });
     cache.set("templates", {
-      main: "<div>{{content}}</div>",
-      recentPost: "<div>{{title}}</div><p>{{author}}</p>"
+      main: {
+        name: "main",
+        extension: "html",
+        path: "",
+        modifiedAt: new Date(),
+        raw: "<div>{{content}}</div>"
+      },
+      recentPost: {
+        name: "recentPost",
+        extension: "html",
+        path: "",
+        modifiedAt: new Date(),
+        raw: "<div>{{title}}</div><p>{{author}}</p>"
+      }
     });
     cache.set(
       "posts",
@@ -209,9 +236,15 @@ describe("Builder - HTML", () => {
   });
 
   test("No archive post template file", async () => {
-    const cache = new MemoryCache();
+    cache.set("config", defaultConfig);
     cache.set("templates", {
-      main: "<div>{{content}}</div>"
+      main: {
+        name: "main",
+        extension: "html",
+        path: "",
+        modifiedAt: new Date(),
+        raw: "<div>{{content}}</div>"
+      }
     });
     cache.set("posts", DirScanner.scanAndGetFiles("./test/data/project/posts"));
 
@@ -229,11 +262,22 @@ describe("Builder - HTML", () => {
   });
 
   test("Archive", async () => {
-    const cache = new MemoryCache();
-    cache.set("config", { postsFolder: "posts" });
+    cache.set("config", { ...defaultConfig, recentPosts: 2 });
     cache.set("templates", {
-      main: "<div>{{content}}</div>",
-      archivePost: "<div>{{title}}</div><p>{{date}}</p>"
+      main: {
+        name: "main",
+        extension: "html",
+        path: "",
+        modifiedAt: new Date(),
+        raw: "<div>{{content}}</div>"
+      },
+      archivePost: {
+        name: "main",
+        extension: "html",
+        path: "",
+        modifiedAt: new Date(),
+        raw: "<div>{{title}}</div><p>{{date}}</p>"
+      }
     });
     cache.set(
       "posts",