+++ test/unit/builder.metadata.test.ts
@@ -15,9 +15,9 @@ describe("Builder - Metadata", () => {
       @b abcdef\n\
       @c 12-12-2019\n\
       @d hi. \"m,y #(n4-m3)!¡?¿ i's Alberto;\n\
-      -->"
+      -->my-content"
     };
-    const { metadata } = getFileMetadata(file);
+    const { metadata, html } = getFileMetadata(file);
     expect(metadata).toMatchObject({
       template: "main",
       a: "1234",
@@ -25,6 +25,7 @@ describe("Builder - Metadata", () => {
       c: "12-12-2019",
       d: "hi. \"m,y #(n4-m3)!¡?¿ i's Alberto;"
     });
+    expect(html).toEqual("my-content");
   });
 
   test("Parse template file (multi line spaces)", () => {
@@ -39,15 +40,29 @@ describe("Builder - Metadata", () => {
       @a 1234\n\
       @b abcdef\n\
       @c 12-12-2019\n\
-      -->"
+      --> my-content"
     };
-    const { metadata } = getFileMetadata(file);
+    const { metadata, html } = getFileMetadata(file);
     expect(metadata).toMatchObject({
       template: "main",
       a: "1234",
       b: "abcdef",
       c: "12-12-2019"
     });
+    expect(html).toEqual(" my-content");
+  });
+
+  test("Parse template file without metadata", () => {
+    const file: IFile = {
+      name: "article",
+      extension: "html",
+      modifiedAt: new Date(),
+      path: "",
+      raw: "my-content"
+    };
+    const { metadata, html } = getFileMetadata(file);
+    expect(metadata).toMatchObject({});
+    expect(html).toEqual("my-content");
   });
 
   test("Parse template file (error not started)", () => {