+++ test/unit/builder.test.ts
@@ -1,10 +1,17 @@
-import { Builder, IBuilders } from "../../src/domain/builder";
+import { Builder, IBuilders, buildHtml } from "../../src/domain/builder";
 import { MemoryCache } from "../../src/lib/cache";
-import { IFile } from "../../src/interfaces";
+import { IFile, ICache } from "../../src/interfaces";
+
+jest.mock("../../src/domain/builder/filetype/html.builder");
 
 describe("Builder", () => {
+  let cache: ICache;
+
+  beforeEach(() => {
+    cache = new MemoryCache();
+  });
+
   test("Build file calls registered builder", async () => {
-    const cache = new MemoryCache();
     const fakeFile: IFile = {
       name: "test",
       extension: "html",
@@ -27,7 +34,6 @@ describe("Builder", () => {
   });
 
   test("Build file calls unregistered builder", async () => {
-    const cache = new MemoryCache();
     const fakeFile: IFile = {
       name: "test",
       extension: "css",
@@ -47,5 +53,34 @@ describe("Builder", () => {
 
     expect(output).toMatchObject(fakeFile);
     expect(fakeBuilder).not.toHaveBeenCalled();
+  });
+
+  test("Build templates", async () => {
+    cache.set("config", { postsFolder: "posts" });
+    cache.set("templates", {
+      main: {
+        name: "main",
+        extension: "html",
+        path: "",
+        modifiedAt: new Date(),
+        raw: "<div>{{content}}</div>"
+      },
+      post: {
+        name: "post",
+        extension: "html",
+        path: "",
+        modifiedAt: new Date(),
+        raw: "<div>{{content}}</div>"
+      }
+    });
+
+    const cacheSetSpy = spyOn(cache, "set");
+    const builder = new Builder(cache, {});
+    await builder.buildTemplates();
+    expect(buildHtml).toHaveBeenCalled();
+    expect(cacheSetSpy).toHaveBeenCalled();
+
+    const templates = cache.get("templates");
+    expect(Object.keys(templates)).toEqual(["main", "post"]);
   });
 });