+++ CHANGELOG.md
@@ -11,11 +11,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Recent posts "Read more" link.
 - Replace post metadata inside post.
+- Templates support using another template (one level only).
 
 ## Changed
 
 - Updated node to 12.13 LTS.
 - File metadata name doesn't include extension.
+- Templates are stored as a IFile.
 
 ## Fixed